--- a/ecmascript/Cargo.toml
+++ b/ecmascript/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "swc_ecmascript"
-version = "0.6.1"
+version = "0.6.2"
 authors = ["강동윤 <kdy1997.dev@gmail.com>"]
 license = "Apache-2.0/MIT"
 repository = "https://github.com/swc-project/swc.git"