--- a/ecmascript/parser/src/parser/util.rs
+++ b/ecmascript/parser/src/parser/util.rs
@@ -1,7 +1,68 @@
 use super::*;
+use crate::token::Keyword;
 use swc_atoms::js_word;
 
 impl Context {
+    pub(crate) fn is_reserved(self, word: &Word) -> bool {
+        match *word {
+            Word::Keyword(Keyword::Let) => self.strict,
+            Word::Keyword(Keyword::Await) => self.in_async || self.strict,
+            Word::Keyword(Keyword::Yield) => self.in_generator || self.strict,
+
+            Word::Null
+            | Word::True
+            | Word::False
+            | Word::Keyword(Keyword::Break)
+            | Word::Keyword(Keyword::Case)
+            | Word::Keyword(Keyword::Catch)
+            | Word::Keyword(Keyword::Continue)
+            | Word::Keyword(Keyword::Debugger)
+            | Word::Keyword(Keyword::Default_)
+            | Word::Keyword(Keyword::Do)
+            | Word::Keyword(Keyword::Export)
+            | Word::Keyword(Keyword::Else)
+            | Word::Keyword(Keyword::Finally)
+            | Word::Keyword(Keyword::For)
+            | Word::Keyword(Keyword::Function)
+            | Word::Keyword(Keyword::If)
+            | Word::Keyword(Keyword::Return)
+            | Word::Keyword(Keyword::Switch)
+            | Word::Keyword(Keyword::Throw)
+            | Word::Keyword(Keyword::Try)
+            | Word::Keyword(Keyword::Var)
+            | Word::Keyword(Keyword::Const)
+            | Word::Keyword(Keyword::While)
+            | Word::Keyword(Keyword::With)
+            | Word::Keyword(Keyword::New)
+            | Word::Keyword(Keyword::This)
+            | Word::Keyword(Keyword::Super)
+            | Word::Keyword(Keyword::Class)
+            | Word::Keyword(Keyword::Extends)
+            | Word::Keyword(Keyword::Import)
+            | Word::Keyword(Keyword::In)
+            | Word::Keyword(Keyword::InstanceOf)
+            | Word::Keyword(Keyword::TypeOf)
+            | Word::Keyword(Keyword::Void)
+            | Word::Keyword(Keyword::Delete) => true,
+
+            // Future reserved word
+            Word::Ident(js_word!("enum")) => true,
+
+            Word::Ident(js_word!("implements"))
+            | Word::Ident(js_word!("package"))
+            | Word::Ident(js_word!("protected"))
+            | Word::Ident(js_word!("interface"))
+            | Word::Ident(js_word!("private"))
+            | Word::Ident(js_word!("public"))
+                if self.strict =>
+            {
+                true
+            }
+
+            _ => false,
+        }
+    }
+
     pub fn is_reserved_word(self, word: &JsWord) -> bool {
         match *word {
             js_word!("let") => self.strict,