## Initial lexer performance
```
test lexer::tests::lex_colors_js               ... bench:      26,520 ns/iter (+/- 9,030) = 43 MB/s
test lexer::tests::lex_colors_ts               ... bench:      26,708 ns/iter (+/- 669) = 43 MB/s
test lexer::tests::lex_dec_lit                 ... bench:      65,231 ns/iter (+/- 970) = 8 MB/s
test lexer::tests::lex_escaped_char            ... bench:       6,966 ns/iter (+/- 128) = 33 MB/s
test lexer::tests::lex_ident                   ... bench:      26,627 ns/iter (+/- 1,253) = 21 MB/s
test lexer::tests::lex_large_number            ... bench:      19,789 ns/iter (+/- 534) = 11 MB/s
test lexer::tests::lex_legact_octal_lit        ... bench:      91,708 ns/iter (+/- 1,362) = 6 MB/s
test lexer::tests::lex_long_ident              ... bench:       4,199 ns/iter (+/- 85) = 113 MB/s
test lexer::tests::lex_regex                   ... bench:       8,643 ns/iter (+/- 144) = 17 MB/s
test lexer::tests::lex_semicolons              ... bench:     135,916 ns/iter (+/- 2,197) = 13 MB/s
```

## After ascii optimization

```
test lexer::tests::lex_colors_js               ... bench:      25,649 ns/iter (+/- 2,109) = 45 MB/s
test lexer::tests::lex_colors_ts               ... bench:      25,876 ns/iter (+/- 237) = 44 MB/s
test lexer::tests::lex_dec_lit                 ... bench:      67,877 ns/iter (+/- 1,130) = 7 MB/s
test lexer::tests::lex_escaped_char            ... bench:       7,576 ns/iter (+/- 279) = 30 MB/s
test lexer::tests::lex_ident                   ... bench:      24,669 ns/iter (+/- 244) = 22 MB/s
test lexer::tests::lex_large_number            ... bench:      20,754 ns/iter (+/- 2,325) = 10 MB/s
test lexer::tests::lex_legact_octal_lit        ... bench:      93,841 ns/iter (+/- 1,305) = 6 MB/s
test lexer::tests::lex_long_ident              ... bench:       4,561 ns/iter (+/- 94) = 104 MB/s
test lexer::tests::lex_regex                   ... bench:       8,611 ns/iter (+/- 1,579) = 18 MB/s
test lexer::tests::lex_semicolons              ... bench:     105,534 ns/iter (+/- 5,907) = 17 MB/s
```

### After removing wrong optimization in `read_word_as_str`

```
test lexer::tests::lex_colors_js               ... bench:      24,661 ns/iter (+/- 264) = 46 MB/s
test lexer::tests::lex_colors_ts               ... bench:      24,712 ns/iter (+/- 523) = 46 MB/s
test lexer::tests::lex_dec_lit                 ... bench:      63,026 ns/iter (+/- 2,060) = 8 MB/s
test lexer::tests::lex_escaped_char            ... bench:       6,949 ns/iter (+/- 889) = 33 MB/s
test lexer::tests::lex_ident                   ... bench:      23,894 ns/iter (+/- 1,355) = 23 MB/s
test lexer::tests::lex_large_number            ... bench:      19,238 ns/iter (+/- 347) = 11 MB/s
test lexer::tests::lex_legact_octal_lit        ... bench:      87,562 ns/iter (+/- 3,667) = 6 MB/s
test lexer::tests::lex_long_ident              ... bench:       4,432 ns/iter (+/- 112) = 107 MB/s
test lexer::tests::lex_regex                   ... bench:       7,945 ns/iter (+/- 394) = 19 MB/s
test lexer::tests::lex_semicolons              ... bench:      96,317 ns/iter (+/- 2,444) = 19 MB/s
```

### After adding inline

```
test lexer::tests::lex_colors_js               ... bench:      24,551 ns/iter (+/- 701) = 47 MB/s
test lexer::tests::lex_colors_ts               ... bench:      24,949 ns/iter (+/- 7,295) = 46 MB/s
test lexer::tests::lex_dec_lit                 ... bench:      62,327 ns/iter (+/- 734) = 8 MB/s
test lexer::tests::lex_escaped_char            ... bench:       7,073 ns/iter (+/- 104) = 32 MB/s
test lexer::tests::lex_ident                   ... bench:      23,956 ns/iter (+/- 481) = 23 MB/s
test lexer::tests::lex_large_number            ... bench:      19,059 ns/iter (+/- 1,107) = 11 MB/s
test lexer::tests::lex_legact_octal_lit        ... bench:      90,779 ns/iter (+/- 3,574) = 6 MB/s
test lexer::tests::lex_long_ident              ... bench:       4,447 ns/iter (+/- 72) = 106 MB/s
test lexer::tests::lex_regex                   ... bench:       8,017 ns/iter (+/- 128) = 19 MB/s
test lexer::tests::lex_semicolons              ... bench:      96,080 ns/iter (+/- 2,390) = 19 MB/s

```

### Baseline for lexer performance

```
test angular       ... bench:  14,032,845 ns/iter (+/- 2,618,442) = 51 MB/s
test backbone      ... bench:   1,845,859 ns/iter (+/- 821,471) = 32 MB/s
test colors        ... bench:      30,071 ns/iter (+/- 8,199) = 38 MB/s
test jquery        ... bench:   9,005,276 ns/iter (+/- 2,257,499) = 29 MB/s
test jquery_mobile ... bench:  14,765,527 ns/iter (+/- 3,800,016) = 30 MB/s
test mootools      ... bench:   7,303,173 ns/iter (+/- 2,145,621) = 21 MB/s
test underscore    ... bench:   1,524,935 ns/iter (+/- 223,030) = 28 MB/s
test yui           ... bench:   8,172,519 ns/iter (+/- 1,928,539) = 41 MB/s
```


### After ascii optimization

I changed some parts of lexer to work with byte instead of char.

```
test angular       ... bench:  10,811,003 ns/iter (+/- 1,482,822) = 66 MB/s
test backbone      ... bench:   1,998,636 ns/iter (+/- 928,390) = 30 MB/s
test colors        ... bench:      26,790 ns/iter (+/- 3,411) = 43 MB/s
test jquery        ... bench:   8,206,829 ns/iter (+/- 1,367,117) = 32 MB/s
test jquery_mobile ... bench:  12,997,307 ns/iter (+/- 1,559,951) = 34 MB/s
test mootools      ... bench:   7,195,759 ns/iter (+/- 2,110,055) = 22 MB/s
test underscore    ... bench:   1,457,760 ns/iter (+/- 13,685) = 29 MB/s
test yui           ... bench:   6,430,216 ns/iter (+/- 1,641,020) = 52 MB/s

```

### After making keyword lexing faster

Previous impl creates `JsWord` and converted it into keyword :)
I don't know why I was such lazy, but I fixed it anyway.

```
test angular       ... bench:   9,571,435 ns/iter (+/- 466,983) = 75 MB/s
test backbone      ... bench:   1,370,984 ns/iter (+/- 61,407) = 43 MB/s
test colors        ... bench:      21,354 ns/iter (+/- 843) = 54 MB/s
test jquery        ... bench:   7,451,739 ns/iter (+/- 299,895) = 36 MB/s
test jquery_mobile ... bench:  12,125,868 ns/iter (+/- 354,864) = 37 MB/s
test mootools      ... bench:   5,902,485 ns/iter (+/- 198,905) = 27 MB/s
test underscore    ... bench:   1,191,789 ns/iter (+/- 61,073) = 36 MB/s
test yui           ... bench:   7,342,329 ns/iter (+/- 102,193) = 46 MB/s
```


### Baseline for parser

```
test angular          ... bench:  28,546,408 ns/iter (+/- 9,048,341) = 25 MB/s
test angular_ts       ... bench:  32,415,280 ns/iter (+/- 4,809,052) = 22 MB/s
test backbone         ... bench:   4,113,307 ns/iter (+/- 656,662) = 14 MB/s
test backbone_ts      ... bench:   4,185,279 ns/iter (+/- 684,611) = 14 MB/s
test colors           ... bench:      72,065 ns/iter (+/- 10,394) = 16 MB/s
test colors_ts        ... bench:      64,925 ns/iter (+/- 1,140) = 17 MB/s
test jquery           ... bench:  23,206,153 ns/iter (+/- 3,700,224) = 11 MB/s
test jquery_mobile    ... bench:  36,288,682 ns/iter (+/- 6,031,766) = 12 MB/s
test jquery_mobile_ts ... bench:  37,104,999 ns/iter (+/- 5,948,459) = 12 MB/s
test jquery_ts        ... bench:  26,216,268 ns/iter (+/- 3,724,721) = 10 MB/s
test large            ... bench:     894,625 ns/iter (+/- 20,068) = 5 MB/s
test mootools         ... bench:  18,456,129 ns/iter (+/- 2,944,629) = 8 MB/s
test mootools_ts      ... bench:  19,072,284 ns/iter (+/- 2,892,291) = 8 MB/s
test underscore       ... bench:   3,787,685 ns/iter (+/- 582,680) = 11 MB/s
test underscore_ts    ... bench:   4,189,735 ns/iter (+/- 604,984) = 10 MB/s
test yui              ... bench:  16,532,152 ns/iter (+/- 2,570,311) = 20 MB/s
test yui_ts           ... bench:  17,706,670 ns/iter (+/- 2,884,282) = 19 MB/s
```

### After optimizing parse_primary_expr

```
test angular          ... bench:  27,067,654 ns/iter (+/- 3,904,212) = 26 MB/s
test angular_ts       ... bench:  29,543,385 ns/iter (+/- 4,093,929) = 24 MB/s
test backbone         ... bench:   3,865,621 ns/iter (+/- 591,432) = 15 MB/s
test backbone_ts      ... bench:   4,322,681 ns/iter (+/- 583,417) = 13 MB/s
test colors           ... bench:      67,511 ns/iter (+/- 782) = 17 MB/s
test colors_ts        ... bench:      69,072 ns/iter (+/- 9,364) = 16 MB/s
test jquery           ... bench:  21,676,669 ns/iter (+/- 3,218,923) = 12 MB/s
test jquery_mobile    ... bench:  35,866,274 ns/iter (+/- 5,232,480) = 12 MB/s
test jquery_mobile_ts ... bench:  35,868,294 ns/iter (+/- 5,154,427) = 12 MB/s
test jquery_ts        ... bench:  22,262,265 ns/iter (+/- 3,387,292) = 12 MB/s
test large            ... bench:     728,133 ns/iter (+/- 5,605) = 6 MB/s
test mootools         ... bench:  18,547,388 ns/iter (+/- 2,571,215) = 8 MB/s
test mootools_ts      ... bench:  20,170,346 ns/iter (+/- 2,803,543) = 7 MB/s
test underscore       ... bench:   3,358,521 ns/iter (+/- 496,402) = 12 MB/s
test underscore_ts    ... bench:   4,011,983 ns/iter (+/- 551,207) = 10 MB/s
test yui              ... bench:  15,861,491 ns/iter (+/- 2,327,605) = 21 MB/s
test yui_ts           ... bench:  16,641,472 ns/iter (+/- 2,630,265) = 20 MB/s
```

### After optimizing skip_space

```
test angular          ... bench:  26,050,513 ns/iter (+/- 5,822,522) = 27 MB/s
test angular_ts       ... bench:  27,010,412 ns/iter (+/- 11,632,935) = 26 MB/s
test backbone         ... bench:   3,350,210 ns/iter (+/- 172,170) = 17 MB/s
test backbone_ts      ... bench:   3,569,779 ns/iter (+/- 165,606) = 16 MB/s
test colors           ... bench:      53,089 ns/iter (+/- 2,680) = 21 MB/s
test colors_ts        ... bench:      54,438 ns/iter (+/- 2,023) = 21 MB/s
test jquery           ... bench:  19,466,370 ns/iter (+/- 668,417) = 13 MB/s
test jquery_mobile    ... bench:  31,501,930 ns/iter (+/- 7,064,350) = 14 MB/s
test jquery_mobile_ts ... bench:  34,467,660 ns/iter (+/- 14,276,566) = 13 MB/s
test jquery_ts        ... bench:  21,525,053 ns/iter (+/- 9,270,297) = 12 MB/s
test large            ... bench:     715,374 ns/iter (+/- 10,540) = 7 MB/s
test mootools         ... bench:  16,998,838 ns/iter (+/- 7,986,295) = 9 MB/s
test mootools_ts      ... bench:  15,614,933 ns/iter (+/- 680,101) = 10 MB/s
test underscore       ... bench:   3,029,955 ns/iter (+/- 135,441) = 14 MB/s
test underscore_ts    ... bench:   3,219,333 ns/iter (+/- 117,416) = 13 MB/s
test yui              ... bench:  14,324,433 ns/iter (+/- 450,292) = 23 MB/s
test yui_ts           ... bench:  18,065,076 ns/iter (+/- 3,243,760) = 18 MB/s
```